requirements.txt
kafka-python>=2.0.2
python-dotenv>=1.0.0
openai>=1.3.0
pydantic>=2.5.0 
watchdog>=3.0.0